move : nextwal 세팅 use case 패키지 이동
