+++ src/main/java/backend/wal/wal/nextwal/application/port/in/GetNextWalUseCase.java
@@ -1,4 +1,4 @@
-package backend.wal.wal.nextwal.application.port.out;
+package backend.wal.wal.nextwal.application.port.in;
 
 import backend.wal.wal.nextwal.domain.NextWals;
 import backend.wal.wal.nextwal.domain.aggregate.NextWal;
