+++ src/main/java/backend/wal/wal/nextwal/application/port/in/NextWalSettingUseCase.java
@@ -1,4 +1,4 @@
-package backend.wal.wal.nextwal.application.port.out;
+package backend.wal.wal.nextwal.application.port.in;
 
 import backend.wal.wal.common.domain.WalCategoryType;
 import backend.wal.wal.nextwal.domain.NextWals;
